added helper class
